refactor(fumigation-report): update report structure to use IndustrialSafetyConditionsDTO
