+++ Backend/anecacao_api/src/main/java/com/anecacao/api/reporting/data/dto/IndustrialSafetyConditionsDTO.java
@@ -1,0 +1,14 @@
+package com.anecacao.api.reporting.data.dto;
+
+import lombok.Data;
+
+@Data
+public class IndustrialSafetyConditionsDTO {
+    private boolean electricDanger;
+    private boolean fallingDanger;
+    private boolean hitDanger;
+
+    public boolean hasAnyDanger() {
+        return electricDanger || fallingDanger || hitDanger;
+    }
+}
